Update emoji-datasource-apple to v15
I noticed that the npm package we use as a source for official unicode
emojis has been updated with v15 emojis [1]. v15 includes things like
:black_bird:, :moose:, :jellyfish:, and :pea_pod: [2]. Important
additions for use on Buildkite!

I followed the docs in `docs/updating-unicode.md` but found they needed
to be tweaked slightly for current yarn versions.

I also notice that one of the rake tasks updated `all_emoji_names.txt`
with the new emojis, but also a range of recently added custom emojis.
Are we supposed to update the file when merging custom emoji PRs?
🤷‍♂️

This is similar to the v14 update in #369, and the v7 update in #330.

[1] https://www.npmjs.com/package/emoji-datasource
[2] https://emojipedia.org/emoji-15.0