avoid conflicting phoenix emojis
Unicode added a phoenix emoji in emoji v15.1, which conflicts with the
logo we add in 2016 (#16) for the elixr framework.

It feels liek it'd be nice to let users access both, but which one wins?
This proposes we let the unicode emoji win, by moving the elixr
framework icon to phoenix-elixr. This will be a visually breaking change for
customers who use this emoji.