remove duplicate aliases from img-buildkite-64.json
When running `rake verify` I noticed that we get some non-fatal
errors:

    $ rake verify
    WARNING: Ignored duplicate emoji names:
     - :cucumber: (cucumber) refers to 2 emoji
     - :llama: (llama) refers to 2 emoji
     - :troll: (troll) refers to 2 emoji
     - :yarn: (yarn) refers to 2 emoji
    ERROR: Unexpected duplicate emoji names:
     - :cucumber-open: (cucumber-open) refers to 2 emoji
     - :esbuild: (esbuild) refers to 2 emoji
     - :garden-io: (garden-io) refers to 2 emoji
     - :gitlab: (gitlab) refers to 2 emoji
     - :kart: (kart) refers to 2 emoji
     - :locust: (locust) refers to 2 emoji
     - :ruff: (ruff) refers to 2 emoji
     - :usertesting: (usertesting) refers to 2 emoji

This fixes all the errors (but not the warnings). In the set of
buildkite custom emoji, we only need to provide value(s) for aliases if
they're different to the canonical name. Somehow we've ended up with
some emojis (like `ruff`) that have an identical alias (`ruff`).